silant/service/serializers.py
from rest_framework import serializers
from .models import *


class MachineSerializer(serializers.ModelSerializer):
    class Meta:
        model = Machine
        fields = ('__all__')

class MaintenanceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Maintenance
        fields = ('__all__')

class ComplaintSerializer(serializers.ModelSerializer):
    class Meta:
        model = Complaint
        fields = ('__all__')
