silant/requirements.txt
asgiref==3.7.2
Django==4.2.1
django-bootstrap5==23.3
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
